Implement SmartLifecycle and improve structure
