+++ functions/supplier/cdc-debezium-supplier2/pom.xml
@@ -19,6 +19,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <version.debezium>2.1.4.Final</version.debezium>
         <spring-cloud.version>2022.0.2</spring-cloud.version>
+        <apicurio.version>2.4.2.Final</apicurio.version>
     </properties>
 
     <dependencyManagement>
@@ -37,13 +38,12 @@
         <dependency>
             <groupId>io.apicurio</groupId>
             <artifactId>apicurio-registry-utils-converter</artifactId>
-            <version>2.4.2.Final</version>
-            <scope>test</scope>
+            <version>${apicurio.version}</version>
         </dependency>
         <dependency>
             <groupId>io.apicurio</groupId>
             <artifactId>apicurio-registry-client</artifactId>
-            <version>2.4.2.Final</version>
+            <version>${apicurio.version}</version>
         </dependency>
 
         <dependency>
@@ -61,12 +61,6 @@
             <groupId>javax.annotation</groupId>
             <artifactId>javax.annotation-api</artifactId>
             <version>1.3.2</version>
-        </dependency>
-
-        <dependency>
-            <groupId>io.debezium</groupId>
-            <artifactId>debezium-core</artifactId>
-            <version>${version.debezium}</version>
         </dependency>
 
         <dependency>
