+++ functions/supplier/cdc-debezium-supplier2/src/main/java/org/springframework/cloud/fn/supplier/cdc/EmbeddedEngineExecutorService.java
@@ -16,34 +16,43 @@
 
 package org.springframework.cloud.fn.supplier.cdc;
 
-import java.io.Closeable;
 import java.io.IOException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import io.debezium.engine.DebeziumEngine;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
+import org.springframework.context.SmartLifecycle;
+
 /**
- *
  * @author Christian Tzolov
  */
-public class EmbeddedEngineExecutorService implements Closeable {
+public class EmbeddedEngineExecutorService implements SmartLifecycle, AutoCloseable {
 
 	private static final Log logger = LogFactory.getLog(EmbeddedEngineExecutorService.class);
 
 	private final DebeziumEngine<?> engine;
 	private final ExecutorService executor;
+	private final AtomicBoolean running = new AtomicBoolean(false);
 
 	public EmbeddedEngineExecutorService(DebeziumEngine<?> engine) {
 		this.engine = engine;
 		this.executor = Executors.newSingleThreadExecutor();
 	}
 
+	@Override
 	public void start() {
 		logger.info("Start Embedded Engine");
 		this.executor.execute(this.engine);
+		this.running.set(true);
+	}
+
+	@Override
+	public void stop() {
+		this.close();
 	}
 
 	@Override
@@ -51,10 +60,16 @@ public class EmbeddedEngineExecutorService implements Closeable {
 		logger.info("Stop Embedded Engine");
 		try {
 			this.engine.close();
+			this.running.set(false);
 		}
 		catch (IOException e) {
 			logger.warn("Failed to close the Debezium Engine:", e);
 		}
 		this.executor.shutdown();
+	}
+
+	@Override
+	public boolean isRunning() {
+		return this.running.get();
 	}
 }
