+++ functions/supplier/cdc-debezium-supplier2/src/test/java/org/springframework/cloud/fn/supplier/cdc/streaming/CdcSourceAvroFormatTest.java
@@ -27,7 +27,7 @@ import org.testcontainers.containers.GenericContainer;
 
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.cloud.fn.supplier.cdc.CdcConfiguration.BindingNameStrategy;
+import org.springframework.cloud.fn.supplier.cdc.BindingNameStrategy;
 import org.springframework.cloud.stream.binder.test.OutputDestination;
 import org.springframework.cloud.stream.binder.test.TestChannelBinderConfiguration;
 import org.springframework.context.ConfigurableApplicationContext;
